exploration/requirements.txt
anthropic>=0.40.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
